project/requirements.txt
asgiref==3.3.1
astroid==2.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-bootstrap-form==3.4
django-import-export==2.4.0
et-xmlfile==1.0.1
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysql-connector==2.2.9
mysqlclient==2.0.2
odfpy==1.4.1
openpyxl==3.0.5
protobuf==3.14.0
pylint==2.6.0
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
